+++ apps/site/src/lib/api/blocks/get.ts
@@ -13,8 +13,8 @@ export const getFeaturedBlocks = async (): Promise<ExpandedBlockMetadata[]> => {
     getDbBlock({ author: "hash", name: "address" }),
     getDbBlock({ author: "hash", name: "how-to" }),
     getDbBlock({ author: "hash", name: "ai-image" }),
-  ]).then((result) =>
-    result.filter((block): block is ExpandedBlockMetadata => !!block),
+  ]).then(
+    (result) => result.filter((block) => !!block) as ExpandedBlockMetadata[],
   );
 };